# Fix stream allocations