@klauspost 2020-05-05
Regression is likely due to 16MB blocks being read instead of 4MB.